Split to two different traits (#187)
